refactor: move CalleeSavedGarbageReadCheck to own file
