+++ riscv_analysis/src/lints/mod.rs
@@ -18,3 +18,6 @@ pub use dead_value::*;
 
 mod garbage_input_value;
 pub use garbage_input_value::*;
+
+mod callee_saved_garbage_read;
+pub use callee_saved_garbage_read::*;
